+++ main/src/main/java/net/oneandone/stool/util/Field.java
@@ -16,7 +16,7 @@
 package net.oneandone.stool.util;
 
 public enum Field implements Info {
-    ID, SELECTED, DIRECTORY, BACKSTAGE, URL, TYPE, BUILDTIME, OWNER, DISK, STATE, UPTIME, CPU, MEM, SERVICE, TOMCAT, DEBUGGER, SUSPEND, JMX, APPS, OTHER;
+    ID, SELECTED, DIRECTORY, BACKSTAGE, URL, TYPE, BUILDTIME, OWNER, CREATOR, DISK, STATE, UPTIME, CPU, MEM, SERVICE, TOMCAT, DEBUGGER, SUSPEND, JMX, APPS, OTHER;
 
     public String toString() {
         return name().toLowerCase();